refactor: Bars 스켈레톤 개선 및 스토리 개선
[#827]
